ViewDeckController swizzles the 'presentModalViewController:animated:' and '' with its own implementations.
These check if a viewdeckcontroller is present on the UIViewcontroller. If it is, it will delegate the presenting of the (modal) view to the viewdeckcontroller instead of the containing controller. This gives better results for rotation and presentation.

If no viewdeckcontroller is present, it just uses the standard implementation.

This also removes the code introduced by a remark of schemer in issue #28, since it's no longer applicable.
I'm not sure if this fixes #16?