Velocity of swipe open/close animations matches gesture speed.
Improved ViewDeck swipe open/close animations so that the velocity
of the animation after swiping now more closely matches the user's
swipe velocity.

Previously, the animation timings were constant and independent of
swipe velocity. Also, on swipe to close, the animation would stutter
at the beginning due to the ease-in timing.