Refactored the init methods of IIViewDeckController, in particular, the initWithCoder: and initWithNibName:bundle: methods.
These methods should chain invoke their super methods. Not doing
so causes some behaviors to break. For example, without chaining
to initWithCoder: in the super class, you can't setup segues in a
storyboard for use with the class, because the segues do not get
created.

The fix is simply to call the appropriate super initializer, and
then call a common method that initializes other properties.